chore: move machine charm test to separate module
